sensai-backend/pyproject.toml
[project]
name = "sensai-backend"
version = "1.0.0"
description = "SensAI Backend API"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "langchain-core==0.3.40",
    "openai==1.109.1",
    "aiohttp==3.12.13",
    "email_validator==2.2.0",
    "python-dotenv==1.0.1",
    "pydantic==2.8.2",
    "pydantic_core==2.20.1",
    "pydantic-settings==2.7.0",
    "backoff==2.2.1",
    "fastapi==0.114.1",
    "uvicorn==0.30.6",
    "boto3==1.37.18",
    "botocore==1.37.18",
    "httpx==0.27.0",
    "imgkit==1.2.3",
    "pydub==0.25.1",
    "pypdf==5.1.0",
    "pdf2image==1.17.0",
    "unidecode==1.3.8",
    "seaborn==0.13.2",
    "aiosqlite==0.21.0",
    "google-auth==2.38.0",
    "pyasn1-modules==0.4.1",
    "apscheduler==3.11.0",
    "httptools==0.6.4",
    "uvloop==0.21.0; sys_platform != 'win32'",
    "watchfiles==1.0.5",
    "websockets==15.0.1",
    "python-multipart==0.0.20",
    "sentry-sdk[fastapi]==2.20.0",
    "google-cloud-bigquery==3.35.0",
    "aiocache==0.12.3",
    "langfuse==3.5.1",
    "jiter==0.8.2",
    "instructor==1.11.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/api"]

[dependency-groups]
dev = [
    "locust==2.32.4",
    "gevent==24.11.1",
    "pre-commit==4.0.1",
    "pytest==8.3.5",
    "pytest-cov==6.1.1",
    "pytest-asyncio==0.26.0",
]
